Feature: transaction history as procedure
