+++ frontend/src/components/dashboard/DashboardTransaction.tsx
@@ -9,12 +9,10 @@ import GasPrice from "../utils/GasPrice";
 import Link from "../utils/Link";
 
 import DashboardTransactionsHistoryChart from "./DashboardTransactionsHistoryChart";
-import {
-  useRecentTransactions,
-  useTransactionHistory,
-} from "../../hooks/subscriptions";
+import { useRecentTransactions } from "../../hooks/subscriptions";
 import { useLatestGasPrice } from "../../hooks/data";
 import { styled } from "../../libraries/styles";
+import { useFetch } from "../../hooks/use-fetch";
 
 const TransactionCardNumber = styled(Row, {
   "& > .col-12": {
@@ -30,8 +28,10 @@ const TransactionCharts = styled(Row, {
 
 const DashboardTransaction: React.FC = React.memo(() => {
   const { t } = useTranslation();
-  const transactionsCountHistoryForTwoWeeks = useTransactionHistory()
-    ?.transactionsCountHistoryForTwoWeeks;
+  const transactionsCountHistoryForTwoWeeks = useFetch(
+    "transaction-history",
+    []
+  );
   const recentTransactionsCount = useRecentTransactions()
     ?.recentTransactionsCount;
   const latestGasPrice = useLatestGasPrice();
