+++ frontend/src/components/dashboard/DashboardTransactionsHistoryChart.tsx
@@ -6,7 +6,7 @@ import moment from "moment";
 import PaginationSpinner from "../utils/PaginationSpinner";
 
 import { useTranslation } from "react-i18next";
-import { TransactionCountHistory } from "../../types/subscriptions";
+import { TransactionCountHistory } from "../../types/procedures";
 
 const chartsStyle = {
   height: 232,
